requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
async-timeout==5.0.1
asynctest==0.13.0
attributee==0.1.9
attrs==24.2.0
blosc2==2.0.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.6.1
Cython==3.0.11
decorator==4.4.2
distro==1.9.0
dotty-dict==1.3.1
einops==0.8.0
exceptiongroup==1.2.2
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.5.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
idna==3.10
imageio==2.4.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iniconfig==2.0.0
interrogate==1.7.0
isort==4.3.21
jiter==0.7.1
kwarray==0.7.1
lap==0.4.0
lapx==0.5.11
mccabe==0.7.0
# mmcls==0.25.0
mmcv-full==1.7.2
# mmdet @ file:///home/jclee/workspace/src/good
# -e git+https://github.com/open-mmlab/mmtracking@e79491ec8f0b8c86fda947fbaaa824c66ab2a991#egg=mmtrack
motmetrics==1.4.0
moviepy==0.2.3.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
numexpr==2.8.6
numpy==1.22.0
onnx==1.7.0
onnxruntime==1.19.2
openai==1.55.0
pandas==1.3.5
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.0
protobuf==3.20.1
py==1.11.0
py-cpuinfo==9.0.0
pybullet==3.2.6
pycocotools==2.0.7
pycodestyle==2.12.1
pydantic==2.10.1
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
pyproject_hooks==1.2.0
pytest==8.3.3
requests==2.32.3
scipy==1.7.3
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
tables==3.8.0
tabulate==0.9.0
terminaltables==3.1.10
tomli==2.0.2
tqdm==4.67.1
typing_extensions==4.12.2
ubelt==1.3.6
urllib3==2.2.3
xdoctest==1.2.0
xmltodict==0.14.2
yapf==0.40.2
yarl==1.15.2
